add a test for overwrite check; re-organize all fixtures into `conftest.py`
